pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = "Simple portfolio calculator"
authors = ["Fabian Hof <Fabian.Hof@pm.me>"]

[tool.poetry.dependencies]
python = "^3.10"
yfinance = "^0.1.63"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = {version = "^21.7b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
